Create a new Logger singleton type, pass to script engine
This is an effort to elliminate the possibility of
openstudio::Singleton<openstudiio::Logger>::instance being non unique
across the ScriptEngine boundary. A new custom Logger singleton has been
introduced that allows us to pass a Logger pointer from the main cli
program into the script engine. This should give us a guarantee that the
Ruby/Python script engines do not spawn their own unique loggers.

This is an experiment and subject to review and refinement.

ref #4847